ASN1: Remove BitStrings and clean up ASN.1/DER conversion
This is very similar to an Uptane project commit accomplishing the
same goal for Uptane's ASN1 code.

Because older code tried to support both BitStrings and OctetStrings,
there was an awkward additional layer of data structures in the ASN.1
definitions, class BinaryData, which could be occupied by either
BitStrings or OctetStrings. BitStrings and BinaryData have been
removed, and anything that is intended to be binary data is now just
a straight OctetString. BitStrings were being encoded as strings
anyway (as opposed to the OctetString, which is encoded as straight
bytes).

In the process, I also simplified the tag constraints on affected
fields. More simplification of other fields will follow in future
commits. These simplifications are cases in which tag constraints
and .subtype() calls didn't seem necessary to me based on my
reading of the ASN.1 documentation, and which, when removed, didn't
seem to make a difference. (Not requiring them also seems to make
the code less sensitive to minor changes, and certainly makes it
more readable. I don't recall at this point, but I think that the
"simplification" here may have also faciliated an update to more
recent pyasn1 versions due to changes in the way they were
interpreted; I am not certain on that last point.)

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>