Fix tuf/tuf#736: Rebuild roledb in refresh()
even if root has not been updated.
In order to avoid freezes if role info is deleted due to prior
validation failures, make sure that roledb is rebuilt during
client tuf.client.updater.refresh(), even if root did not need
to be updated.

If, for example, snapshot.json is deleted because timestamp has
indicated that there is a newer version but that newer version
cannot be obtained, we should still not fail to try to update
snapshot next time.
See https://github.com/theupdateframework/tuf/issue/736

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>