allow a TUF client to override the system clock
to some value that they trust and will manually update.

This is of use if, say, a trusted time is periodically retrieved
from some other source, and the clock will stand still until it
is updated this way.  For example, Uptane has use for this
functionality.

Edit: also includes two minor comment corrections suggested by
Lukas.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>