Move expiration checks for Root to the right place
Root metadata must be verified even if it is expired.  This is
important for root chaining.  It is only necessary to check
whether or not Root is expired when we are done updating Root,
and we want to update the other top-level roles.  At that point,
if our currently-trusted Root metadata is expired, we MUST stop
and only continue once we have unexpired Root metadata.

Consequently, this commit:
- kills Root expiration checks in _verify_uncompressed_metadata_file
- adds a Root expiration check in refresh(), before we try to
  update the other top-level roles
- adds some TODOs suggesting other expiration checks to exclude
- adds a TODO suggesting the migration to the new specification's
  update-root-first-and-don't-check-root-against-snapshot-hashes
  model.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>